+++ base-compat/CHANGES.markdown
@@ -12,6 +12,7 @@
    later, as that is the first version of `base` to provide the `traceEvent`
    primitive on which `traceEventWith` is defined.
  - Backport `inits1` and `tails1` to `Data.List.NonEmpty.Compat`.
+ - Backport `minusNaturalMaybe` to `Numeric.Natural.Compat`.
 
 ## Changes in 0.12.2 [2022.08.11]
  - Sync with `base-4.17`/GHC 9.4